Move allocation worker functions to worker.c and remove unnecessary peer_id arg from cdt_allocate_shared_page
